+++ spec/lib/facebook_ads/ad_objects/server_side/event_request_spec.rb
@@ -16,6 +16,7 @@
 # IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 # CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+require 'json'
 
 RSpec.describe 'FacebookAds::ServerSide::EventRequest' do
     it 'constructor works' do
@@ -106,5 +107,63 @@ RSpec.describe 'FacebookAds::ServerSide::EventRequest' do
         )
         expect(event_request1).to_not eq(event_request2)
         expect(event_request1.hash).to_not eq(event_request2.hash)
+    end
+
+    it 'execute works when all fields are specified' do
+        event = FacebookAds::ServerSide::Event.new(event_name: 'event-1')
+        expected_params = {
+            data: [JSON.generate(event.normalize)],
+            test_event_code: 'test_event_code-1',
+            partner_agent: 'partner_agent-2',
+            namespace_id: 'namespace_id-3',
+            upload_id: 'upload_id-4',
+            upload_tag: 'upload_tag-5',
+            upload_source: 'upload_id-6',
+        }
+        event_request = FacebookAds::ServerSide::EventRequest.new(
+            pixel_id: expected_params[:pixel_id],
+            events: [event],
+            test_event_code: expected_params[:test_event_code],
+            partner_agent: expected_params[:partner_agent],
+            namespace_id: expected_params[:namespace_id],
+            upload_id: expected_params[:upload_id],
+            upload_tag: expected_params[:upload_tag],
+            upload_source: expected_params[:upload_source],
+        )
+        mock_ads_pixel = double('mock-ads-pixel')
+        response = {
+            events_received: ['e1', 'e2'],
+            messages: 'message3',
+            fbtrace_id: 'fbtrace123',
+        }
+        expected_response = FacebookAds::ServerSide::EventResponse.new(response)
+        expect(mock_ads_pixel).to receive_message_chain('events.create').with(expected_params).and_return(response)
+        expect(FacebookAds::AdsPixel).to receive(:get).with(event_request.pixel_id).and_return(mock_ads_pixel)
+        response = event_request.execute
+
+        expect(response).to eq(expected_response)
+    end
+
+    it 'execute works when only events are specified' do
+        event = FacebookAds::ServerSide::Event.new(event_name: 'event-1')
+        expected_params = {
+            data: [JSON.generate(event.normalize)],
+        }
+        event_request = FacebookAds::ServerSide::EventRequest.new(
+            pixel_id: expected_params[:pixel_id],
+            events: [event],
+        )
+        mock_ads_pixel = double('mock-ads-pixel')
+        response = {
+            events_received: ['e1', 'e2'],
+            messages: 'message3',
+            fbtrace_id: 'fbtrace123',
+        }
+        expected_response = FacebookAds::ServerSide::EventResponse.new(response)
+        expect(mock_ads_pixel).to receive_message_chain('events.create').with(expected_params).and_return(response)
+        expect(FacebookAds::AdsPixel).to receive(:get).with(event_request.pixel_id).and_return(mock_ads_pixel)
+        response = event_request.execute
+
+        expect(response).to eq(expected_response)
     end
 end